Extracts table into own component, starts Progress component, adds icons to nav
