+++ web/src/routes/index.tsx
@@ -1,8 +1,9 @@
 import { component$, useContext } from '@builder.io/qwik';
 import { type DocumentHead } from "@builder.io/qwik-city";
 
-import Hero from "../components/furniture/hero";
-import SectionLinkGrid from "../components/psc/section-link-grid";
+import Hero from "~/components/furniture/hero";
+import SectionLinkGrid from "~/components/psc/section-link-grid";
+import Progress from "~/components/psc/progress";
 
 import { ChecklistContext } from '~/store/checklist-context';
 
@@ -12,6 +13,7 @@ export default component$(() => {
   return (
     <>
       <Hero />
+      <Progress />
       <SectionLinkGrid sections={checklists.value} />
     </>
   );
